Move jobserver func into job_token module
